refactor the old patches module
deduplicate code between ovos-core and workshop

support get_response refactor
